feat: Remove duplicate StationaryLight creation.
